requirements.txt
opencv-python
numpy
tqdm
pillow